JAVA-2754 move DefaultQuarkusCqlSession to a dedicated session package
